remindir/requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-recurrence==1.9.0
django-registration==3.0.1
django-rest-auth==0.9.5
django-webpack-loader==0.6.0
djangorestframework==3.9.4
idna==2.8
oauthlib==3.0.1
pycodestyle==2.5.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
